+++ src/screens/planItemActivity/PlanItemTaskScreen.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
 
-import { Card, FullScreenTemplate, Icon, IconButton, IconToggleButton, StyledText } from 'components';
+import { FullScreenTemplate } from 'components';
 import { i18n } from 'locale';
 import { PlanItem, PlanItemType } from 'models';
 import { NavigationInjectedProps } from 'react-navigation';
-import { dimensions, getElevation, palette, typography } from 'styles';
-import { PlanItemComplexTask } from './PlanItemComplexTask';
-import { PlanItemSimpleTask } from './PlanItemSimpleTask';
+import { PlanItemForm, PlanItemFormData } from './PlanItemForm';
 
 interface State {
   planItem: PlanItem;
@@ -20,10 +17,8 @@ export class PlanItemTaskScreen extends React.PureComponent<NavigationInjectedPr
   };
 
   state: State = {
-    planItem: this.props.navigation.getParam('planItem') ? this.props.navigation.getParam('planItem') : new PlanItem(),
-    taskType: this.props.navigation.getParam('planItem')
-      ? this.props.navigation.getParam('planItem').type
-      : PlanItemType.SimpleTask,
+    planItem: this.props.navigation.getParam('planItem'),
+    taskType: PlanItemType.SimpleTask,
   };
 
   handleChangeText = (name: string) => {
@@ -38,93 +33,31 @@ export class PlanItemTaskScreen extends React.PureComponent<NavigationInjectedPr
     });
   };
 
+  createPlanItem = async (name: string) => {
+    const plan = this.props.navigation.getParam('plan');
+
+    const planItem = await PlanItem.createPlanItem(plan, PlanItemType.SimpleTask, name);
+
+    this.setState({ planItem });
+  };
+
+  updatePlanItem = async (name: string) => {
+    await this.state.planItem.update({
+      name,
+    });
+
+    this.setState({ planItem: { ...this.state.planItem, name } });
+  };
+
+  onSubmit = ({ name }: PlanItemFormData) =>
+    this.state.planItem ? this.updatePlanItem(name) : this.createPlanItem(name);
+
   render() {
-    const { planItem, taskType } = this.state;
+    const { planItem } = this.state;
     return (
       <FullScreenTemplate darkBackground>
-        <View style={styles.subHeaderContainer}>
-          <View>
-            <StyledText style={styles.textInput}>{i18n.t('planItemActivity:taskNamePlaceholder')}</StyledText>
-          </View>
-          <View style={styles.buttonsContainer}>
-            <IconToggleButton icons={['layers-clear', 'layers']} onPress={this.changePlanItemType} />
-            <IconButton
-              name="mic-off"
-              type="material"
-              size={24}
-              color={palette.primaryVariant}
-              containerStyle={styles.iconButtonContainer}
-            />
-          </View>
-        </View>
-        <Card style={styles.card}>
-          {taskType === PlanItemType.SimpleTask ? (
-            <PlanItemSimpleTask planItem={planItem} />
-          ) : (
-            <PlanItemComplexTask planItem={planItem} />
-          )}
-        </Card>
+        <PlanItemForm planItem={planItem} onSubmit={this.onSubmit} />
       </FullScreenTemplate>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  subHeaderContainer: {
-    height: 56,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: dimensions.spacingExtraLarge,
-    backgroundColor: palette.background,
-    ...getElevation(5),
-    borderBottomColor: palette.backgroundAdditional,
-    borderBottomWidth: 1,
-  },
-  buttonsContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  textInput: {
-    ...typography.subtitle,
-    color: palette.textInputPlaceholder,
-  },
-  card: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'flex-start',
-    marginVertical: dimensions.spacingBig,
-    marginHorizontal: dimensions.spacingExtraLarge,
-    height: '78%',
-  },
-  iconButtonContainer: {
-    backgroundColor: palette.backgroundAdditional,
-    paddingVertical: 4,
-    paddingHorizontal: dimensions.spacingSmall,
-    borderRadius: 8,
-  },
-  imagePicker: {
-    borderRadius: 8,
-    borderColor: palette.backgroundSurface,
-    borderWidth: 1,
-    display: 'flex',
-    paddingHorizontal: 91,
-    paddingVertical: 67,
-  },
-  imagePickerContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: dimensions.spacingSmall,
-  },
-  imageInputText: {
-    ...typography.taskInput,
-    color: palette.textInputPlaceholder,
-    marginTop: 53,
-  },
-  timerButton: {
-    position: 'absolute',
-    right: dimensions.spacingBig,
-    top: dimensions.spacingBig,
-  },
-});
